change spin table double to int

+++ api/database/migrations/2025_03_11_142755_create_spin_list_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name')->nullable();
             $table->unsignedTinyInteger('status')->default(1);  
-            $table->decimal('prize', 10, 2)->nullable();  
+            $table->int('prize')->nullable();  
             $table->timestamps();
         });
     }
